requirements-new.txt
# Core audio processing dependencies
numpy==1.23.5
scipy==1.9.3
librosa==0.8.1
soundfile==0.13.1
pydub==0.25.1

# Machine learning
tensorflow==2.12.1
scikit-learn==1.2.2

# Audio source separation
spleeter==2.4.2

# Data processing
pandas==1.5.3
matplotlib==3.6.3

# Configuration management
PyYAML==6.0.1

# HTTP client for services
requests==2.31.0

# Development dependencies (optional)
pytest==7.4.2
pytest-cov==4.1.0
black==23.7.0
flake8==6.0.0
mypy==1.5.1

# Documentation (optional)
sphinx==7.1.2
sphinx-rtd-theme==1.3.0